lib: troubleshoot image loading
Even after shoving all the computation into preemptive threads, it seems like
`ImagePNG.parsefile` is locking the GUI for some reason. Or maybe it's locking
Lwt or Lwt_react.

My guess now is the `chunk_reader` has some state, and `parsefile` is messing with it
in a way that Lwt doesn't like. But it's weird because the file does process. But log
output from "Converted ..." to "ran_color ..." blocks while clustering is processing.

Maybe I will try CamlImages for JPG file loading.